+++ CHANGELOG.md
@@ -1,7 +1,13 @@
 # Change Log
 All notable changes to this project will be documented in this file. This change log follows the conventions of [keepachangelog.com](http://keepachangelog.com/).
 
-## [Unreleased]
-### Changed
+## 1.0.0
+### Added
+
+- parse-str->Document : Parse xml string into JDOM Document
+- parse-str :  Parse xml string into JDOM Document and return the root element
+- find-first : Return the first element at the XPath
+- find-all : Return collection of elements at the XPath
+- get-value : Return the value at the XPath as a String, regardless of whether the XPath points to a element or attribute